COM fixes

+++ VK9-Library/CCubeTexture9.h
@@ -42,7 +42,7 @@ public:
 	D3DPOOL mPool;
 	HANDLE* mSharedHandle;
 
-	int mReferenceCount;
+	ULONG mReferenceCount;
 	VkResult mResult;
 
 public:
