+++ VK9-Library/CIndexBuffer9.h
@@ -38,13 +38,13 @@ public:
 	CIndexBuffer9(CDevice9* device, UINT Length, DWORD Usage, D3DFORMAT Format, D3DPOOL Pool, HANDLE* pSharedHandle);
 	~CIndexBuffer9();
 
-	int mReferenceCount;
+	ULONG mReferenceCount;
 	VkResult mResult;
 	void* mData;
 	size_t mSize;
 	size_t mCapacity;
 	bool mIsDirty;
-	int mLockCount;
+	uint32_t mLockCount;
 
 	VkMemoryRequirements mMemoryRequirements;
 	VkBuffer mBuffer;
